% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{calculate_years_since_onset}
\alias{calculate_years_since_onset}
\title{Calculate years since disease onset}
\usage{
calculate_years_since_onset(
  df,
  age_col = "age",
  onset_age_col = "onset_age",
  new_col = "years_since_onset"
)
}
\arguments{
\item{df}{A data.frame/tibble.}

\item{age_col}{Column name for age. Default: \code{"age"}.}

\item{onset_age_col}{Column name for onset age. Default: \code{"onset_age"}.}

\item{new_col}{Name of the new column to create. Default: \code{"years_since_onset"}.}
}
\value{
A tibble with \code{new_col} added as \code{age - onset_age}.
}
\description{
Calculate years since disease onset
}
